Output all Prism parser errors (#9660)
We previously only outputted Prism parser errors when there was a
missing node. This was because we were concerned about too much noise
when there was still so much work to do on integrating Prism. However,
now that we are getting closer to feature parity with the original
parser, it would be useful to output the full set of Prism parser
errors.

A lot of the time, invalid Ruby code results in a missing node, but
that's not always the case. For example:

```ruby
module
  def foo; end
end
```

Running this with `--parser=prism --stop-after=desugarer` will currently
show no errors, because we don't hit the missing node case. Instead, we
will now see:

    test.rb:2: unexpected constant path after `module`; class/module name must be CONSTANT https://srb.help/2001
         2 |  def foo; end
                       ^^^

The equivalent message for the original parser:

    test.rb:2: unexpected token tNL https://srb.help/2001
         2 |  def foo; end
         3 |end

These aren't identical, and in some cases we may find the Prism error is
less helpful than the original parser (or vice versa). We're planning
follow up work to ensure the error recovery and messages for Prism
parser mode are good, but for now this commit just enables outputting
all the parser errors.