Fix crash with unterminated regular expressions in Prism parser mode
The Prism parser was causing a crash with code containing unterminated
regular expression literals:

```ruby
/foo
```

Prism's error recovery includes the rest of the file content in the
regex pattern:

```
@ InterpolatedRegularExpressionNode (location: (1,0)-(1,4))
├── flags: newline, static_literal
├── opening_loc: (1,0)-(1,1) = "/"
├── parts: (length: 1)
│   └── @ StringNode (location: (1,1)-(1,4))
│       ├── flags: static_literal, frozen
│       ├── opening_loc: ∅
│       ├── content_loc: (1,1)-(1,4) = "foo"
│       ├── closing_loc: ∅
│       └── unescaped: "foo"
└── closing_loc: (1,4)-(1,4) = ""
```

The `closing_loc` has has both start and end at the same position. When
translating regexp options, we increment the `closingLoc.start` by 1 to
skip past the closing regex delimiter. However, in the case of an
unterminated regex there is no closing deliter, so the start offset is
greater than the end offset, causing an assertion failure.

This commit fixes the issue by using the original 0 length location and
only incrementing the start offset if appropriate.

Original desugar tree:

    class <emptyTree><<C <root>>> < (::<todo sym>)
      ::Regexp.new("", 0)
    end

Prism desugar tree:

    class <emptyTree><<C <root>>> < (::<todo sym>)
      ::Regexp.new("foo\n# typed: true", 0)
    end

The resulting behavior is different from the original parser, since the
remainder of the file is captured as part of the regex rather than
discarded. As such, we're skipping verifying the exp trees match the
original parser outputs.