`--single_test=file` → `--single_test file`
Copy the underlying test runner invocation out of a failing bazel test
target, which looks something like this:

    ==================== Test output for //test:test_LSPTests/testdata/lsp/fast_path/shuffle_and_change:
    + exec test/lsp_test_runner --single_test test/testdata/lsp/fast_path/shuffle_and_change.rb

Then I paste this into the terminal, prefix it with
`lldb -- bazel-bin/`, and that's how I quickly run a test under the
debugger.

Sometimes, I want to run a different test. It's a little easier for tab
completion by editing a previous `lldb --` command from my history. But
when I do that, I don't get tab completion, because there is no special
tab completion for the `lldb --` that would let it know that what
follows after `--single_test` is a filename.

So instead, let's just use a space, so that tab completion works.