Add another test
The first test added in this PR crashed, because there was an
index-out-of-bounds.

This test does not actually crash: it just demonstrates incorrect
behavior.

What would happen is that the `idx` of the owner for the
`FoundDefinitionRef` for the `on_b` method in the old
`FoundDefinitionHashes` is 2 in (`[<root>, A, B]`).

But when you reinterpret that index into the list of defined classes in
the new file, the thing at index 2 is `A` (`[<root>, A, A, B]`). That
means we attempt to delete methods with name `on_b` from `A`, of which
there are none, and fail to delete the methods from `B`. That then looks
like a method redefinition error, because a second `on_b(x)` definition
comes in but we haven't cleared out the old `on_b()` definition.