Rewrite using inheritance to avoid bloating 4 bytes
At first, when the assertion failed that we were storing an extra 8
bytes on the FileHash, I was like, "yep, I expected this, let's change
it." But then when I looked where the error was, the CheckSize assertion
was on a struct I wasn't expecting: it was on the `FileHash`, when I was
naively expecting it was on the `FoundDefHash` or something.

So then before adding, I went to look at what we were storing on the
`FileHash` and whether there were already any spare padding bytes.

There were not on any of the top-level fields, but there **were** on the
`LocalSymbolTableHashes`, which was only using 52 of its 58 bytes. So
then I spent a bunch of time trying to figure out how to get my new
field into those 4 bytes of an ~unrelated structure: the
`LocalSymbolTableHashes` is what's produced by `GlobalState::hash`,
which is what my new `uint32_t` is used to initialize (e.g., the new
`hash(...)` parameter).

But then I thought, wait, what's the point of trying to put my int in
that padding: we're *already* storing every bit of it that we need by
`mix`'ing it into the `hierarchyHash` of the `LocalSymbolTableHashes`.
We don't actually read the new `foundClassesHash` once it got embedded
into the `FileHash` struct: it only needs to be around in memory, not
persisted.

So at that point, I opted to chagne the out parameter. We'll still
store `FoundDefHashes` in the `FileHash`, but we'll **compute** a
subclass of that `FoundDefHashes` which has extra fields that are not
meant to be persisted. I opted to use inheritance instead of composition
for this because I didn't want to have to refactor all of the usage
sites to indirect through a new member name, but if people think the
inheritance approach is too clever, I am happy to back that out and use
composition.