Make `void` be a module, not a class
Realistically, we should treat `void` methods as returning, essentially, some abstract type.

Sorbet doesn't have abstract types, and the best that we can get is to return `T.anything`, the top type.

But if we return the top type, then we lose the ability to associate custom error messages for improper usages of `void`-typed things.

Instead, we say that `void` methods return
`Sorbet::Private::Static::Void`, which is a `ClassType` backed by a
`class`. Unfortunately, this means that `void`-typed things gain the
same `ClassType` subtyping relationships that entails, so `void` is a
subtype of `Object` and `BasicObject`.

This is pretty rare, because most of the time when people misuse
`void`-typed method results, it's to call methods on that result. But
Sorbet already has a specific error message for that.