Have to move the typeExpr later
It's read by `getResultType` to populate the `type` field on the `Cast`,
so we can't move it until after we've populated the `type` field.

This also makes for a convenient way to determine whether we need to do
something with the tree in builder_walk to keep for IDE requests, vs
having to reimplement the `isLambdaTLet` function.