Move `T.let(->(){}, ...)` rewrite to resolver (#9739)
* Move `T.let(->(){}, ...)` rewrite to resolver

The tree is mutable in resolver, and will remain so ~indefinitely.

By contrast, I'd like to get to a point where we no longer mutate the
trees after resolver, so that we can cache them and then reuse them for
queries like hover and go to def.

* Have to move the typeExpr later

It's read by `getResultType` to populate the `type` field on the `Cast`,
so we can't move it until after we've populated the `type` field.

This also makes for a convenient way to determine whether we need to do
something with the tree in builder_walk to keep for IDE requests, vs
having to reimplement the `isLambdaTLet` function.

* Actually, use EmptyTree

It doesn't look like we call `sanityCheck` on the tree after resolver,
but `sanityCheck` **does** assert that `typeExpr` is not `nullptr`, so I
figure it's probably better to not use `nullptr`. We can put `EmptyTree`
there, to say that it was explicit.

* Add an ENFORCE