Make parser option cache sensitive (#9752)
* Make parser option cache sensitive

This commit ensures we do not share the cache between original and prism
parser modes. I noticed while testing that running `tc` first with
`--parser=prism` changed the output of `tc` when running with
`--parser=original`. Given the two parsers produce ever so slightly
different desugared representations of the same source code, this is
not surprising.

Although we'd like for the desugared representation to be identical, in
some cases we may opt to produce a different AST, e.g. in error recovery
cases where the parsers work quite differently. At least for now, it
makes sense to not share the cache between the two parser modes.

* Remove parser option in favor of cache sensitive flag