Remove `getIndexed`
I'm working towards making it so that we cache the resolved trees in LSP
mode, instead of the indexed trees.

Right now we have this `getIndexed` API in LSPTypechecker that
represents a "public API" for interacting with the type checker (e.g.
LSP requests can use this API freely to implement language-aware tasks).

Currently, there are no uses of this API. I thought that `getResolved`
used it, but it looks like it depends directly on the vectors storing
indexed trees directly.

Given there are no users, I'd like to remove it so that we don't get any
new ones cropping up accidentally while I work towards that larger
change.