Fix crash with changed class definitions in a file (#9728)
* Add failing test

* Move this test into fast_path folder

* Move populateFoundDefHashes to SymbolDefiner

* clang-format

* Fix failing test with more hashing

* Add another test

The first test added in this PR crashed, because there was an
index-out-of-bounds.

This test does not actually crash: it just demonstrates incorrect
behavior.

What would happen is that the `idx` of the owner for the
`FoundDefinitionRef` for the `on_b` method in the old
`FoundDefinitionHashes` is 2 in (`[<root>, A, B]`).

But when you reinterpret that index into the list of defined classes in
the new file, the thing at index 2 is `A` (`[<root>, A, A, B]`). That
means we attempt to delete methods with name `on_b` from `A`, of which
there are none, and fail to delete the methods from `B`. That then looks
like a method redefinition error, because a second `on_b(x)` definition
comes in but we haven't cleared out the old `on_b()` definition.

* Rewrite using inheritance to avoid bloating 4 bytes

At first, when the assertion failed that we were storing an extra 8
bytes on the FileHash, I was like, "yep, I expected this, let's change
it." But then when I looked where the error was, the CheckSize assertion
was on a struct I wasn't expecting: it was on the `FileHash`, when I was
naively expecting it was on the `FoundDefHash` or something.

So then before adding, I went to look at what we were storing on the
`FileHash` and whether there were already any spare padding bytes.

There were not on any of the top-level fields, but there **were** on the
`LocalSymbolTableHashes`, which was only using 52 of its 58 bytes. So
then I spent a bunch of time trying to figure out how to get my new
field into those 4 bytes of an ~unrelated structure: the
`LocalSymbolTableHashes` is what's produced by `GlobalState::hash`,
which is what my new `uint32_t` is used to initialize (e.g., the new
`hash(...)` parameter).

But then I thought, wait, what's the point of trying to put my int in
that padding: we're *already* storing every bit of it that we need by
`mix`'ing it into the `hierarchyHash` of the `LocalSymbolTableHashes`.
We don't actually read the new `foundClassesHash` once it got embedded
into the `FileHash` struct: it only needs to be around in memory, not
persisted.

So at that point, I opted to chagne the out parameter. We'll still
store `FoundDefHashes` in the `FileHash`, but we'll **compute** a
subclass of that `FoundDefHashes` which has extra fields that are not
meant to be persisted. I opted to use inheritance instead of composition
for this because I didn't want to have to refactor all of the usage
sites to indirect through a new member name, but if people think the
inheritance approach is too clever, I am happy to back that out and use
composition.

* Clarify a comment

* Fix namer_test

* I think the `idx` is redundant

* Update infer_test

* Fix pipeline_test_runner