ClassDef should return last element of class (#9778)
* ClassDef should return last element of class

This doesn't actually matter for us, because we don't allow putting a
sig on a `<static-init>` and thus you can't actually use the result
of evaluating a class def.

But I'm working on a change that removes class_flatten, and it's easier
to model the "true" Ruby behavior where the last thing is returned, than
it is to model the current behavior. So for the sake of showing the
change to the exp files, I want to make this as a separate change.

* ga '**/*.flatten-tree.exp'

* ga '**/*.flatten-tree-raw.exp'

* ga '**/*.cfg-text.exp'

* Use different name for exp file

* Two more