Fix printing for CFGText, AST, and ASTRaw
There are two ways things can be printed:

- to stdout
- to a file specified by the `-p`/`--print` flag

By not calling `fmt` on the right PrinterConfig, we were not making the
print output show up in the right file, if such a file was specified.

I noticed that this was happening on #9777, but @elliottt actually took
the time to figure out the cause.

Going to land this as its own change.